% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.heatmap.r
\name{sig_heatmap}
\alias{sig_heatmap}
\title{sig_heatmap}
\usage{
sig_heatmap(all_patients_result, title)
}
\arguments{
\item{all_patients_result}{result table}

\item{title}{title}
}
\value{
pheatmap
}
\description{
create significant heatmap
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
